requirements/base.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.11.2
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
django-tinymce==2.8.0
django-jazzmin